+++ pyvista/core/composite.py
@@ -14,7 +14,7 @@ from vtk import vtkMultiBlockDataSet
 import pyvista
 from pyvista.utilities import get_array, is_pyvista_dataset, wrap
 
-from .common import DataObject
+from .dataset import DataObject
 from .filters import CompositeFilters
 
 log = logging.getLogger(__name__)
